backend/requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart
google-generativeai
json_repair
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
pydantic[email]
pydantic-settings
Pillow
pytesseract
pymupdf
pymupdf4llm
